A6: split AudienceBookingRepository into Reader + Writer behind a façade (#563 backlog)
AudienceBookingReader holds the 13 read queries (lookups, conflict detection,
affected-user collection, count); AudienceBookingWriter holds the 10 writes
(create/update/cancel/delete + the N:N booking↔audience / booking↔user
mutators) plus the private FOR UPDATE conflict helper, which belongs to the
write transaction. Both share the same cache group; the create() transaction
(START TRANSACTION / FOR UPDATE / COMMIT-ROLLBACK) is preserved verbatim.
AudienceBookingRepository stays the public façade with the @phpstan-type and a
tech-debt note (#563 B3).

Behavior-preserving — AudienceBookingRepositoryTest green (95 tests);
PHPStan L8 + WPCS clean.
